@@ -1,36 +1,54 @@
-'''
-Image accusition using Micro-Manager's Python (2) bindings (Camera class)
-and a server program (CameraServer class).
+'''The camera server for using the camera and save images.
 
-On Windows, MM builds come compiled with Python 2 support only, so in this solution
-there is a Python 2 server program that controls the camera and image saving
-and then the client end that can be run with Python 3.
+Functionality
+-------------
+- Image acquisition using Micro-Manager's Python bindings (pymmcore)
+- Live feed using matplotlib
+- File saving using the tifffile module
+
+
+Background
+-----------
+On Windows, MM builds came precompiled with Python 2 support only. The
+camera server/client division here made it possible to run the server
+calling MM using Python 2 (to control the camera and image saving) and
+then the camera client run with Python 3.
+
+It is still usefull; Because it uses (network) sockets, it is trivial
+to use another (or multiple PC) to acqurie images. Or, in future, to
+use another backends than MM.
 '''
 
 import os
+import sys
 import time
 import datetime
 import socket
+import argparse
 import threading
 import multiprocessing
 
-import MMCorePy
+try:
+    import pymmcore
+except ImportError:
+    pymmcore = None
+    print('pymmcore not installed')
 import tifffile
 import numpy as np
 import matplotlib.pyplot as plt
 from matplotlib.animation import FuncAnimation
 from matplotlib.widgets import RectangleSelector
 
-import camera_communication as cac
-from camera_communication import SAVING_DRIVE
+from .camera_communication import PORT
 
-DEFAULT_SAVING_DIRECTORY = "D:\imaging_data"
+DEFAULT_SAVING_DIRECTORY = "imaging_data"
+DEFAULT_MICROMANAGER_DIR = 'C:/Program Files/Micro-Manager-2.0'
 
-
+# Integer between 1-inf (1 = no downsampling), images for imageshower
+LIVE_DOWNSAMPLE = 2
 
 class ImageShower:
-    '''
-    Showing images on the screen on its own window.
+    '''Shows images on the screen in its own window.
 
     In future, may be used to select ROIs as well to allow
     higher frame rate imaging / less data.
@@ -45,7 +63,6 @@ class ImageShower:
     Methods
     -------
     self.loop       Set this as multiprocessing target
-
     '''
     def __init__(self):
         self.fig = plt.figure()
@@ -95,7 +112,7 @@ class ImageShower:
             data = self.queue.get(True, timeout=0.01)
         if data is None:
             return self.im, ''
-        elif data == 'close':
+        elif isinstance(data, str) and data == 'close':
             self.close = True
             return self.im, ''
 
@@ -115,14 +132,13 @@ class ImageShower:
         
         
         per95 = np.percentile(inspect_area, 95)
-        data = np.clip(data, np.percentile(inspect_area, 5), per95)
+        per5 = np.percentile(inspect_area, 5)
+        data = np.clip(data, per5, per95)
         
-        data = data - np.min(data)
-        data_max = np.max(data)
-        data = data.astype(float)/data_max
+        data -= per5
+        data /= (per95-per5)
 
         self.image_size = data.size
-       
         
         self.im.set_array(data)
         self.fig.suptitle('Selection 95th percentile: {}'.format(per95), fontsize=10)
@@ -130,7 +146,7 @@ class ImageShower:
         return self.im, text
            
          
-    def loop(self, queue):
+    def loop(self, queue, title):
         '''
         Runs the ImageShower by reading images from the given queue.
         Set this as a multiprocessing target.
@@ -142,53 +158,94 @@ class ImageShower:
         
         image = queue.get()
         self.im = plt.imshow(1000*image/np.max(image), cmap='gray', vmin=0, vmax=1, interpolation='none', aspect='auto')
-        self.ani = FuncAnimation(plt.gcf(), self._updateImage, frames=range(100), interval=5, blit=False)
+        self.ani = FuncAnimation(plt.gcf(), self._updateImage, frames=range(100), interval=50, blit=False)
 
-        plt.show(block=False)
+        self.fig.canvas.toolbar.winfo_toplevel().title(title)
+
+        # Remove the toolbar; Gives more space when having many cameras
+        self.fig.canvas.toolbar.pack_forget()
         
-        while not self.close:
-            plt.pause(1)
+        # Take away all white space
+        plt.subplots_adjust(top=1, bottom=0, right=1,left=0,
+                            hspace=0, wspace=0)
+        
+        plt.show(block=True)
 
 
 class DummyCamera:
+    '''A dummy camera suitable for testing the server/client.
     '''
-    A dummy camera class, used when unable to load the real Camera class
-    due to camera being off or something similar.
-    '''
+    def __init__(self):
+        self.settings = {'setting1' : 'na', 'setting2': 0.0, 'setting3': 1}
+        self.camera = None
+
     def acquire_single(self, save, subdir):
         pass
-    def acquire_series(self, exposure_time, image_interval, N_frames, label, subdir, trigger_direction):
+    def acquire_series(self, exposure_time, image_interval, N_frames, label, subdir):
         pass
     def save_images(images, label, metadata, savedir):
         pass
     def set_saving_directory(self, saving_directory):
         pass
     def set_binning(self, binning):
-        pass
+        self.settings['binning'] = binning
+    def set_roi(self, x,y,w,h):
+        self.settings['roi'] = [x,y,w,h]
+    def set_save_stack(self, boolean):
+        self.settings['save-stack'] = boolean
+
     def save_description(self, filename, string):
         pass
     def close(self):
         pass
+    def get_cameras(self):
+        return ['dummy1', 'dummy2']
+    def get_camera(self):
+        return self.camera
+    def set_camera(self, name):
+        self.camera = name
+    def get_settings(self):
+        return list(self.settings.keys())
+    def get_setting_type(self, setting_name):
+        if setting_name == 'setting1':
+            return 'string'
+        elif setting_name == 'setting2':
+            return 'float'
+        elif setting_name == 'setting3':
+            return 'integer'
+        else:
+            print('Invalid setting')
+            return ''
+    def get_setting(self, setting_name):
+        value = self.settings.get(setting_name, None)
+        if value is None:
+            return ''
+        return value
+    def set_setting(self, setting_name, value):
+        self.settings[setting_name] = value
 
 
-class Camera:
-    '''
-    Controlling ORCA FLASH 4.0 camera using Micro-Manager's
-    Python (2) bindings.
+
+class MMCamera:
+    '''Controls any camera using MicroManager and its pymmcore bindings.
     '''
 
     def __init__(self, saving_directory=DEFAULT_SAVING_DIRECTORY):
 
         self.set_saving_directory(saving_directory)
         
-        self.mmc = MMCorePy.CMMCore() 
-        self.mmc.loadDevice('Camera', 'HamamatsuHam', 'HamamatsuHam_DCAM')
-        self.mmc.initializeAllDevices()
-        self.mmc.setCameraDevice('Camera')
+        self.mmc = pymmcore.CMMCore() 
+        self.mmc.setDeviceAdapterSearchPaths([DEFAULT_MICROMANAGER_DIR])
+
+        self._device_name = None
+        self._configuration_name = ''
+
+        #self.mmc.loadDevice('Camera', 'HamamatsuHam', 'HamamatsuHam_DCAM')
+        #self.mmc.initializeAllDevices()
+        #self.mmc.setCameraDevice('Camera')
             
-        self.settings = {'binning': '1x1'}
+        self.settings = {'exposure_time_scaler': 1}
         
-        self.mmc.prepareSequenceAcquisition('Camera')
         #self.mmc.setCircularBufferMemoryFootprint(4000)
         self.live_queue= False
 
@@ -199,9 +256,101 @@ class Camera:
 
         self.save_stack = False
 
+        self.title = 'Camera not set'
+        self.servertitle = ''
 
 
-    def acquire_single(self, save, subdir):
+    def get_cameras(self):
+        '''Lists available MicroManager configuration files in .
+        '''
+        return [fn for fn in os.listdir(DEFAULT_MICROMANAGER_DIR) if fn.endswith('.cfg')]
+
+
+    def get_camera(self):
+        '''Returns the label of the current camera.
+        '''
+        return self._configuration_name
+
+    def set_camera(self, name):
+        '''Set the provided configuration file.
+        '''
+        if not os.path.exists(name):
+            path = os.path.join(DEFAULT_MICROMANAGER_DIR, name)
+            
+            if not os.path.exists(path):
+                print(f'Couldnt open configuration file named: {path}')
+                return
+
+        self.mmc.loadSystemConfiguration(path)
+        
+        self._device_name = self.mmc.getCameraDevice()
+        self._configuration_name = name
+        self.mmc.prepareSequenceAcquisition(self._device_name)
+
+        self.title = f'{name} ({self._device_name}) | {self.servertitle}'
+
+    def get_settings(self):
+        '''Returns device property names
+        '''
+        if self._device_name is None:
+            return ''
+        properties = self.mmc.getDevicePropertyNames(self._device_name)
+        
+        return list(properties) + list(self.settings.keys())
+
+
+    def get_setting_type(self, setting_name):
+        '''Returns "string", "integer" or "float".
+
+        Returns an empty string if the setting does not exist.
+        '''
+        if setting_name in self.settings:
+            return 'float'
+
+        try:
+            num = self.mmc.getPropertyType(self._device_name, setting_name)
+        except RuntimeError as e:
+            print(f'Error! No setting named: {setting_name}')
+            return ''
+        
+        if num == 1:
+            return 'string'
+        elif num == 2:
+            return 'float'
+        elif num == 3:
+            return 'integer'
+
+
+    def get_setting(self, setting_name):
+        if setting_name in self.settings:
+            return self.settings[setting_name]
+        return self.mmc.getProperty(self._device_name, setting_name)
+
+    def set_setting(self, setting_name, value):
+        
+        # a) Internal setting
+        if setting_name in self.settings:
+            self.settings[setting_name] = value
+            return
+        
+        # b) Camera device setting
+        type_name = self.get_setting_type(setting_name)
+        if type_name == 'float':
+            value = float(value)
+        elif type_name == 'integer':
+            value = int(value)
+        elif type_name == '':
+            print('Error! No setting named: {setting_name}')
+            return
+
+        print(f'Changing {setting_name} to its new value {value}')
+        try:
+            self.mmc.setProperty(self._device_name, setting_name, value)
+        except RuntimeError as e:
+            print('Error! The set value likely out of range.')
+            print(e)
+
+    def acquire_single(self, exposure_time, save, subdir):
         '''
         Acquire a single image.
 
@@ -209,11 +358,11 @@ class Camera:
         subdir      Subdirectory for saving
         '''
         
-        exposure_time = 0.01
-        binning = '2x2'
-
-        self.set_binning(binning)
+        exposure_time = float(exposure_time)
         self.mmc.setExposure(exposure_time*1000)
+        #binning = '2x2'
+
+        #self.set_binning(binning)
 
         start_time = str(datetime.datetime.now())
  
@@ -222,22 +371,25 @@ class Camera:
         
         if not self.live_queue:
             self.live_queue = multiprocessing.Queue()
-            self.live_queue.put(image)
+            self.live_queue.put(
+                    image[0::LIVE_DOWNSAMPLE, 0::LIVE_DOWNSAMPLE])
             
-            self.livep = multiprocessing.Process(target=self.shower.loop, args=(self.live_queue,))
+            self.livep = multiprocessing.Process(
+                    target=self.shower.loop,
+                    args=(self.live_queue,self.title))
             self.livep.start()
             
-        self.live_queue.put(image)
+        self.live_queue.put(image[0::LIVE_DOWNSAMPLE, 0::LIVE_DOWNSAMPLE])
 
         if save == 'True':
-            metadata = {'exposure_time_s': exposure_time, 'binning': binning, 'function': 'acquireSingle', 'start_time': start_time}
+            metadata = {'exposure_time_s': exposure_time, 'function': 'acquireSingle', 'start_time': start_time}
 
             save_thread = threading.Thread(target=self.save_images,args=([image],'snap_{}'.format(start_time.replace(':','.').replace(' ','_')), metadata,os.path.join(self.saving_directory, subdir)))
             save_thread.start()
 
 
 
-    def acquire_series(self, exposure_time, image_interval, N_frames, label, subdir, trigger_direction):
+    def acquire_series(self, exposure_time, image_interval, N_frames, label, subdir):
         '''
         Acquire a series of images
 
@@ -246,7 +398,6 @@ class Camera:
         N_frames            How many images to take
         label               Label for saving the images (part of the filename later)
         subdir
-        trigger_direction   "send" (camera sends a trigger pulse when it's ready) or "receive" (camera takes an image for every trigger pulse)
         '''
 
         exposure_time = float(exposure_time)
@@ -254,42 +405,51 @@ class Camera:
         N_frames = int(N_frames)
         label = str(label)
 
-        print "Now aquire_series with label " + label
-        print "- IMAGING PARAMETERS -"
-        print " exposure time " + str(exposure_time) + " seconds"
-        print " image interval " + str(image_interval) + " seconds"
-        print " N_frames " + str(N_frames)
-        print "- CAMERA SETTINGS"
+        print("Now aquire_series with label " + label)
+        print("- IMAGING PARAMETERS -")
+        print(" exposure time " + str(exposure_time) + " seconds")
+        print(" image interval " + str(image_interval) + " seconds")
+        print(" N_frames " + str(N_frames))
+        print("- CAMERA SETTINGS")
 
-        self.set_binning('2x2')
-        print " Pixel binning 2x2"
+        #self.set_binning('2x2')
+        #print(" Pixel binning 2x2")
+        
+        device_name = self.mmc.getDeviceName(self._device_name)
+        print(f'Device name {device_name}')
 
-        if trigger_direction == 'send':
-            print " Camera sending a trigger pulse"
-            self.mmc.setProperty('Camera', "OUTPUT TRIGGER KIND[0]","EXPOSURE")
-            self.mmc.setProperty('Camera', "OUTPUT TRIGGER POLARITY[0]","NEGATIVE")
-        elif trigger_direction== 'receive':
-            print " Camera recieving / waiting for a trigger pulse"
-            self.mmc.setProperty('Camera', "TRIGGER SOURCE","EXTERNAL")
-            self.mmc.setProperty('Camera', "TriggerPolarity","POSITIVE")
-        else:
-            raise ValueError('trigger_direction has to be {} or {}, not {}'.format('receive', 'send', trigger_direction))
+        #if 'hamamatsu' in device_name.lower():
+        #    if trigger_direction == 'send':
+        #        print(" Camera sending a trigger pulse")
+        #        self.mmc.setProperty(self._device_name, "OUTPUT TRIGGER KIND[0]","EXPOSURE")
+        #        self.mmc.setProperty(self._device_name, "OUTPUT TRIGGER POLARITY[0]","NEGATIVE")
+        #    elif trigger_direction== 'receive':
+        #        print(" Camera recieving / waiting for a trigger pulse")
+        #        self.mmc.setProperty(self._device_name, "TRIGGER SOURCE","EXTERNAL")
+        #        self.mmc.setProperty(self._device_name, "TriggerPolarity","POSITIVE")
+        #    elif trigger_direction == 'none':
+        #        pass
+        #    else:
+        #        raise ValueError('trigger_direction has to be send, receive or none, not {trigger_direction}')
 
         
-        print "Circular buffer " + str(self.mmc.getCircularBufferMemoryFootprint()) + " MB"
+        print("Circular buffer " + str(self.mmc.getCircularBufferMemoryFootprint()) + " MB")
 
-        self.mmc.setExposure(exposure_time*1000)
+        scaler = float(self.settings['exposure_time_scaler'])
+        exposure = scaler*exposure_time*1000
+        self.mmc.setExposure(exposure)
 
-
+        self.mmc.clearCircularBuffer()
+        #self.mmc.prepareSequenceAcquisition(self._device_name)
         self.wait_for_client()
         
-
         start_time = str(datetime.datetime.now())
-        self.mmc.startSequenceAcquisition(N_frames, image_interval, False)
-        
+        self.mmc.startSequenceAcquisition(N_frames, image_interval+(1-scaler)*exposure, False)
         
         while self.mmc.isSequenceRunning():
-            time.sleep(exposure_time)
+            self.mmc.sleep(1000*exposure_time)
+
+        self.mmc.sleep(1000)
 
         images = []
 
@@ -298,9 +458,10 @@ class Camera:
                 try:
                     image = self.mmc.popNextImage()
                     break
-                except MMCorePy.CMMError:
-                    time.sleep(exposure_time)
-                
+                except:
+                    # Index error for example when circular buffer is still empty
+                    self.mmc.sleep(1000*exposure_time)
+            
             images.append(image)
             
             
@@ -311,7 +472,9 @@ class Camera:
         save_thread = threading.Thread(target=self.save_images, args=(images,label,metadata,os.path.join(self.saving_directory, subdir)))
         save_thread.start()
         
-        self.mmc.setProperty('Camera', "TRIGGER SOURCE","INTERNAL")
+        #if 'hamamatsu' in device_name.lower() and trigger_direction == 'receive':
+        #    self.mmc.setProperty(self._device_name, "TRIGGER SOURCE","INTERNAL")
+        
         print('acquired')
 
     
@@ -320,17 +483,22 @@ class Camera:
         Save given images as grayscale tiff images.
         '''
         if not os.path.isdir(savedir):
-            os.makedirs(savedir)
+            try:
+                os.makedirs(savedir)
+            except:
+                # May fail if many local servers creating
+                # the folders simultaneously
+                pass
 
         if self.save_stack == False:
             # Save separate images
             for i, image in enumerate(images):
                 fn = '{}_{}.tiff'.format(label, i)
-                tifffile.imsave(os.path.join(savedir, fn), image, metadata=metadata)
+                tifffile.imwrite(os.path.join(savedir, fn), image, metadata=metadata)
         else:
             # Save a stack
             fn = '{}_stack.tiff'.format(label)
-            tifffile.imsave(os.path.join(savedir, fn), np.asarray(images), metadata=metadata)
+            tifffile.imwrite(os.path.join(savedir, fn), np.asarray(images), metadata=metadata)
         
         self.save_description(os.path.join(savedir, 'description'), self.description_string, internal=True)
 
@@ -340,7 +508,6 @@ class Camera:
         Sets where the specimen folders are saved and if the directory
         does not yet exist, creates it.
         '''
-        saving_directory = os.path.join(SAVING_DRIVE, saving_directory)
         if not os.path.isdir(saving_directory):
             os.makedirs(saving_directory)
             
@@ -363,7 +530,7 @@ class Camera:
         Binning '2x2' for example.
         '''
         if not self.settings['binning'] == binning:
-            self.mmc.setProperty('Camera', 'Binning', binning)
+            self.mmc.setProperty(self._device_name, 'Binning', binning)
             self.settings['binning'] =  binning
 
     def set_roi(self, x,y,w,h):
@@ -400,49 +567,55 @@ class Camera:
         
         # Check if the folder exists
         if not os.path.exists(os.path.dirname(fn)):
-            #raise OSError('File {} already exsits'.format(fn))
-            os.makedirs(os.path.dirname(fn))
+            try:
+                os.makedirs(os.path.dirname(fn), exist_ok=True)
+            except:
+                # My fail if many local servers. Let's not
+                # care about it, another server has made it
+                pass
         
-        with open(fn+'.txt', 'w') as fp:
-            fp.write(desc_string)
+        try:
+            with open(fn+'.txt', 'w') as fp:
+                fp.write(desc_string)
 
-        print "Wrote file " + fn+'.txt'
-        
+            print("Wrote file " + fn+'.txt') 
+        except:
+            pass
         
         self.description_string = desc_string
 
 
     def close(self):
-        self.live_queue.put('close')
-        self.lifep.join()
+        if self.live_queue:
+            self.live_queue.put('close')
 
     def wait_for_client(self):
         pass
 
 
 class CameraServer:
+    '''Camera server listens incoming connections from the client and
+    controls a camera class.
     '''
-    Camera server listens incoming connections and
-    controls the camera through Camera class
-    '''
-    def __init__(self):
 
-        PORT = cac.PORT
+    def __init__(self, camera, port=None):
+        
+        if port is None:
+            port = PORT
         HOST = ''           # This is not cac.SERVER_HOSTNAME, leave empty
 
         self.running = False
 
-        self.socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        self.socket.bind((HOST, PORT))
-        self.socket.listen(1)
+        print(f'Binding a socket (host {HOST}, port {port}))')
 
-        try:
-            self.cam = Camera()
-            self.cam.wait_for_client = self.wait_for_client
-        except Exception as e:
-            print e
-            print "Using DUMMY camera instead"
-            self.cam = DummyCamera()
+        self.socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+        self.socket.bind((HOST, port))
+        self.socket.listen(1)
+        
+        print(f'Using the camera <{camera.__class__.__name__}>')
+        self.cam = camera
+        self.cam.servertitle = f'Server on port {port}'
+        self.cam.wait_for_client = self.wait_for_client
         
         self.functions = {'acquireSeries': self.cam.acquire_series,
                           'setSavingDirectory': self.cam.set_saving_directory,
@@ -450,28 +623,37 @@ class CameraServer:
                           'saveDescription': self.cam.save_description,
                           'set_roi': self.cam.set_roi,
                           'set_save_stack': self.cam.set_save_stack,
+                          'get_cameras': self.cam.get_cameras,
+                          'get_camera': self.cam.get_camera,
+                          'set_camera': self.cam.set_camera,
+                          'get_settings': self.cam.get_settings,
+                          'get_setting_type': self.cam.get_setting_type,
+                          'get_setting': self.cam.get_setting,
+                          'set_setting': self.cam.set_setting,
                           'ping': self.ping,
                           'exit': self.stop}
 
-    def ping(self, message):
-        print message
+        self.responding = set([
+            'get_cameras', 'get_camera', 'get_settings', 'get_setting_type', 'get_setting'])
 
+
+    def ping(self, message):
+        print(message)
 
 
     def wait_for_client(self):
-        '''
-        Waits until client confirms that it is ready
+        '''Waits until client confirms that it is ready by sending us
+        anything (usually ping).
         '''
         conn, addr = self.socket.accept()
         string = ''
         while True:
             data = conn.recv(1024)
             if not data: break
-            string += data
+            string += data.decode()
         conn.close()
-        print "Client ready"
+        print("Client ready")
         
-
 
     def run(self):
         '''
@@ -479,27 +661,55 @@ class CameraServer:
         Each established connection can give one command and then the connection
         is closed.
         '''
+        
+        print('Waiting for clients to connect')
+
         self.running = True
         while self.running:
             conn, addr = self.socket.accept()
             string = ''
             while True:
                 data = conn.recv(1024)
-                if not data: break
-                string += data
+                #if not data: break
+                string += data.decode()
+                break
+
+            if not string:
+                conn.close()
+                continue
             
-            conn.close()
             print('Recieved command "'+string+'" at time '+str(time.time()))
-            if string:
+            if ';' in string:
                 func, parameters = string.split(';')
                 parameters = parameters.split(':')
-                target=self.functions[func](*parameters)
+            else:
+                func = string
+                parameters = None
+        
+            # Can close connection early, no response so let's not delay the client
+            if not func in self.responding:
+                conn.close()
             
+            if parameters:
+                response = self.functions[func](*parameters)
+            else:
+                response = self.functions[func]()
+
+            # Say back the response and close because still open
+            if func in self.responding:
+                
+                if isinstance(response, (list, tuple)):
+                    response = ':'.join(response)
+
+                conn.sendall(str(response).encode())
+                conn.close()
+
+
     def stop(self, placeholder):
         '''
         Stop running the camera server.
         '''
-        self.camera.close()
+        self.cam.close()
         self.running = False
 
 
@@ -513,13 +723,41 @@ def test_camera():
 
 
 
-def run_server():
-    '''
-    Running the server.
-    '''
-    cam_server = CameraServer()
+def main():
+    
+    parser = argparse.ArgumentParser(
+            prog='GonioImsoft Camera Server',
+            description='Controls a MicroManager camera')
+
+    parser.add_argument('-p', '--port')
+    parser.add_argument('-c', '--camera')
+    parser.add_argument('-s', '--save-directory')
+
+    args = parser.parse_args()
+
+
+    if args.camera == 'mm':
+        Camera = MMCamera
+    elif args.camera == 'dummy':
+        Camera = DummyCamera
+    else:
+        # Default
+        if pymmcore:
+            Camera = MMCamera
+        else:
+            Camera = DummyCamera
+
+    camera = Camera()
+
+    if args.save_directory:
+        camera.set_saving_directory(args.save_directory)
+
+    if args.port:
+        args.port = int(args.port)
+
+    cam_server = CameraServer(camera, args.port)
     cam_server.run()
             
         
 if __name__ == "__main__":
-    run_server()
+    main()
