@@ -1,18 +1,7 @@
-'''
-Common settings for camera_sever.py (Python2) and camera_client.py (Python3)
+'''Common settings shared by the camera server/client
 '''
 
-# Hostname (or IP address) of the server. This is the address that the client
-# tries to connect to
+# Hostname (or IP address) of the server. This is the address that the
+# client tries to connect to
 SERVER_HOSTNAME = '127.0.0.1'
-
 PORT = 50075
-
-# If for some reason the default port is in use / does not work,
-# try in order some other from this list
-# FIXME NOT IMPLEMENTED YET
-# ALTERNATIVE_PORTS = list(range(50071, 50080))
-
-
-
-SAVING_DRIVE = 'D:\\'
