@@ -1,5 +1,10 @@
-'''
-Places where to save and load temporary files.
+'''Places for saving and loading files.
+
+Constants
+---------
+USERDATA_DIR : string
+    Path to the user data that contains small files
+    such as settings.
 '''
 
 import os
@@ -9,6 +14,27 @@ CODE_ROOTDIR = os.path.dirname(os.path.realpath(__file__))
 USER_HOMEDIR = os.path.expanduser('~')
 
 if platform.system() == "Windows":
-    PUPILDIR = os.path.join(USER_HOMEDIR, 'GonioImsoft')
+    USERDATA_DIR = os.path.join(USER_HOMEDIR, 'GonioImsoft')
 else:
-    PUPILDIR = os.path.join(USER_HOMEDIR, '.gonioimsoft')
+    USERDATA_DIR = os.path.join(USER_HOMEDIR, '.gonioimsoft')
+
+if os.path.isdir(USERDATA_DIR):
+    IS_USERDATA_INITIALIZED = True
+else:
+    IS_USERDATA_INITIALIZED = False
+
+def initialize_userdata():
+    '''Create all user data diretories used by GonioImsoft
+
+    It is good to ask for the user's consent before running this
+    function unless USERDATA_DIR exists already.
+    '''
+    dirs = [
+            USERDATA_DIR,
+            os.path.join(USERDATA_DIR, 'macros'),
+            os.path.join(USERDATA_DIR, 'biosyst_stimuli'),
+            os.path.join(USERDATA_DIR, 'presets'),
+            os.path.join(USERDATA_DIR, 'camera_states'),
+            ]
+    for adir in dirs:
+        os.makedirs(adir, exist_ok=True)
